refactor: no need to store contacts on message
With small amounts of data it isn't really a problem, however, when the
list of contacts starts to grow in size storing it on the message is
unnecessary and can cause performance issues.
If the information about what message was sent to what contacts is
required this can be retrieved from the recipts.